Changed code structure, readers/v2.rs is wip
