+++ src/models/version.rs
@@ -1,0 +1,9 @@
+#[derive(Debug)]
+pub struct Version {
+
+    pub tag: String,
+    pub major: u8,
+    pub minor: u8,
+    pub revision: u8
+
+}
